requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.35.0
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
eval_type_backport==0.2.2
exceptiongroup==1.2.2
feedgen==1.0.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
invoke==2.2.0
isodate==0.7.2
jsonpath-python==1.0.6
lxml==5.3.0
mistralai==1.9.10
mypy-extensions==1.0.0
primp==0.9.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
